+++ crates/infra/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased (YYYY-MM-DD)
 
+- Added `skip_codepoints()`, a non-allocating alternative of `collect_codepoints()`.
+
 ## 0.2.2 (2023-11-08)
 
 - Fixed typos in documentation for [`trim_collapse_ascii_whitespace()`] and [`is_surrogate_utf16()`]